perf: refine manage settings cache boundaries
