+++ apps/app/src/app/(app)/(pending-not-allowed)/[slug]/(workspace)/(manage)/settings/_components/team-general-settings-client.tsx
@@ -16,20 +16,16 @@ import {
 } from "@repo/ui/components/ui/form";
 import { Input } from "@repo/ui/components/ui/input";
 import { toast } from "@repo/ui/components/ui/sonner";
-import {
-  useMutation,
-  useQueryClient,
-  useSuspenseQuery,
-} from "@tanstack/react-query";
+import { useSuspenseQuery } from "@tanstack/react-query";
 import { useOrganizationList } from "@vendor/clerk";
 import { Loader2 } from "lucide-react";
 import { useRouter } from "next/navigation";
-import { useCallback, useMemo, useState } from "react";
+import { useCallback, useMemo } from "react";
 import { useTRPC } from "~/trpc/react";
 import {
   normalizeTeamSlugInput,
-  renameOrganizationSlug,
-} from "./team-general-settings-model";
+  useTeamNameUpdate,
+} from "./team-general-settings-actions";
 
 interface TeamGeneralSettingsClientProps {
   slug: string;
@@ -40,18 +36,10 @@ export function TeamGeneralSettingsClient({
 }: TeamGeneralSettingsClientProps) {
   const router = useRouter();
   const trpc = useTRPC();
-  const queryClient = useQueryClient();
   const { setActive } = useOrganizationList();
-  const [isUpdating, setIsUpdating] = useState(false);
-
-  const orgListQueryOptions = useMemo(
-    () => trpc.viewer.organization.listUserOrganizations.queryOptions(),
-    [trpc]
-  );
-  const orgListQueryKey = orgListQueryOptions.queryKey;
 
   const { data: organizations } = useSuspenseQuery({
-    ...orgListQueryOptions,
+    ...trpc.viewer.organization.listUserOrganizations.queryOptions(),
     staleTime: 5 * 60 * 1000,
   });
   const currentOrg = useMemo(
@@ -70,57 +58,25 @@ export function TeamGeneralSettingsClient({
   const currentFormName = form.watch("teamName");
   const hasChanges = currentFormName !== slug;
 
-  // Optimistic cache update so sidebar and header reflect the new name instantly
-  const updateNameMutation = useMutation(
-    trpc.org.settings.organization.updateName.mutationOptions({
-      meta: { errorTitle: "Failed to update team name" },
-
-      onMutate: async (input) => {
-        await queryClient.cancelQueries({ queryKey: orgListQueryKey });
-        const previousOrgs = queryClient.getQueryData(orgListQueryKey);
-        queryClient.setQueryData(orgListQueryKey, (old: typeof previousOrgs) =>
-          renameOrganizationSlug(old, input)
-        );
-        return { previousOrgs };
-      },
-
-      onError: (_err, _input, context) => {
-        if (context?.previousOrgs) {
-          queryClient.setQueryData(orgListQueryKey, context.previousOrgs);
+  const handleTeamUpdated = useCallback(
+    async (data: { id: string; name: string }) => {
+      try {
+        if (setActive) {
+          await setActive({ organization: data.id });
         }
-      },
 
-      onSuccess: async (data) => {
-        toast.success("Team updated!", {
-          description: `Team name changed to "${data.name}"`,
-        });
-
-        // Update Clerk's active organization before navigation
-        // This ensures cookies are updated before the RSC request
-        try {
-          if (setActive) {
-            await setActive({ organization: data.id });
-          }
-
-          // Refresh router cache to clear any stale RSC data
-          router.refresh();
-
-          // Navigate with updated cookies (client-side navigation)
-          router.push(`/${data.name}/settings`);
-        } catch (error) {
-          console.error("Failed to set active organization:", error);
-          // Still navigate, but may cause temporary mismatch
-          router.push(`/${data.name}/settings`);
-        }
-      },
-
-      onSettled: () => {
-        // Invalidate to ensure consistency with server
-        void queryClient.invalidateQueries({ queryKey: orgListQueryKey });
-        setIsUpdating(false);
-      },
-    })
+        router.refresh();
+        router.push(`/${data.name}/settings`);
+      } catch (error) {
+        console.error("Failed to set active organization:", error);
+        router.push(`/${data.name}/settings`);
+      }
+    },
+    [router, setActive]
   );
+  const { isUpdating, updateTeamName } = useTeamNameUpdate({
+    onUpdated: handleTeamUpdated,
+  });
 
   const onSubmit = useCallback(
     async (values: TeamSettingsFormValues) => {
@@ -133,14 +89,9 @@ export function TeamGeneralSettingsClient({
         return;
       }
 
-      setIsUpdating(true);
-
-      updateNameMutation.mutate({
-        slug,
-        name: values.teamName,
-      });
+      updateTeamName(values.teamName, slug);
     },
-    [form, slug, updateNameMutation.mutate]
+    [form, slug, updateTeamName]
   );
 
   const handleTeamNameChange = useCallback(
