+++ apps/app/src/app/(app)/(pending-not-allowed)/[slug]/(workspace)/(manage)/settings/billing/_components/billing-overview-actions.ts
@@ -1,0 +1,18 @@
+"use client";
+
+import { useQueryClient } from "@tanstack/react-query";
+import { useCallback } from "react";
+import { useTRPC } from "~/trpc/react";
+
+export function useBillingOverviewRefresh() {
+  const trpc = useTRPC();
+  const queryClient = useQueryClient();
+
+  return useCallback(
+    () =>
+      queryClient.invalidateQueries(
+        trpc.org.settings.orgBilling.overview.queryFilter()
+      ),
+    [queryClient, trpc]
+  );
+}
