+++ apps/app/src/app/(app)/(pending-not-allowed)/[slug]/(workspace)/(manage)/settings/billing/page.tsx
@@ -4,9 +4,7 @@ import { BillingSettingsClient } from "./_components/billing-settings-client";
 export const dynamic = "force-dynamic";
 
 export default function BillingPage() {
-  const overviewQueryOptions =
-    trpc.org.settings.orgBilling.overview.queryOptions();
-  prefetch(overviewQueryOptions);
+  prefetch(trpc.org.settings.orgBilling.overview.queryOptions());
 
   return (
     <HydrateClient>
